Update README.md
The autistic mind can get you so far, but a procrastinator's mind will not reach far enough as I will struggle to find the time and place to succeed. But do not worry, I will not disappoint you all as drawing is my passion and will help me carry forward to my future.